Split polling and change tracking
