+++ src/Bot/Features/ChangeTracking/Core/Extensions/ServiceCollectionExtensions.cs
@@ -1,0 +1,15 @@
+using Microsoft.Extensions.DependencyInjection;
+using Mira.Features.ChangeTracking.Core.Repositories;
+
+namespace Mira.Features.ChangeTracking.Core;
+
+public static class ServiceCollectionExtensions
+{
+    public static void AddChangeTracking(this IServiceCollection services)
+    {
+        services.AddTransient<StreamStatusService>();
+        services.AddTransient<BroadcastBoxClient>();
+        services.AddTransient<QueryRepository>();
+        services.AddTransient<CommandRepository>();
+    }
+}
