+++ src/Bot/Features/ChangeTracking/Core/Models/StreamRecord.cs
@@ -1,6 +1,6 @@
 using Mira.Features.Shared;
 
-namespace Mira.Features.Polling.Models;
+namespace Mira.Features.ChangeTracking.Core.Models;
 
 public class StreamRecord
 {
