+++ src/Bot/Features/Polling/PollingService.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Mira.Features.ChangeTracking.Core;
 using Mira.Features.Polling.Options;
 using Mira.Features.Polling.Repositories;
 using Host = Mira.Features.Polling.Models.Host;
