+++ src/Bot/Features/Polling/ServiceCollectionExtensions.cs
@@ -9,10 +9,7 @@ public static class ServiceCollectionExtensions
     public static void AddPollingService(this IServiceCollection services)
     {
         services.AddHostedService<PollingService>(); // TODO: Set this to ignore failures
-        services.AddTransient<StreamStatusService>();
-        services.AddTransient<BroadcastBoxClient>();
-        services.AddTransient<QueryRepository>();
-        services.AddTransient<CommandRepository>();
+
         
         services
             .AddOptions<PollingOptions>()
